@@ -13,7 +13,7 @@ var Environment = require('./environment');
 var NavigatableMixin = {
 
   contextTypes: {
-    router: React.PropTypes.component,
+    router: React.PropTypes.any,
   },
 
   /**
