@@ -17,7 +17,7 @@ var RouterMixin = {
   },
 
   childContextTypes: {
-    router: React.PropTypes.component
+    router: React.PropTypes.any
   },
 
   getChildContext: function() {
@@ -27,7 +27,7 @@ var RouterMixin = {
   },
 
   contextTypes: {
-    router: React.PropTypes.component
+    router: React.PropTypes.any
   },
 
   getInitialState: function() {
